Add getter functions to the APIs
I didn't take the time to research if there's a better way than a full
query and loop to get a specific Did/Name information, but this
currently works.

I wanted to create an application that used this library to toggle a
bulb state without keeping my own state.  I wanted to still be able to
use the phone app and not confuse the state in my pygop-enabled
application so querying the gateway state was a better solution.